Defining an apartness relation on a local ring (#4664)
* Add ring apartness to iset.mm

This is the syntax #r and df-apr .  This has no especially close
analogue in set.mm.

* Add apartness predicate symbol to iset.mm

This is similar to the existing predicate for tight apartness.

* Add dftapap to iset.mm

This connects Ap and TAp in the obvious way.

* Add aptap to iset.mm

Connects two definitions of apartness.

* Add aprirr and aprsym to iset.mm

These are two of the three apartness properties.

Irreflexivity, aprirr , requires that the ring has to be a nonzero ring.

Symmetry is aprsym .

Both of them follow fairly directly from theorems we already have.

Includes aprval which expands the df-apr definition.

* Add local rings to iset.mm

This is the syntax LRing and df-lring .

* Move dvrdir and rdivmuldivd from mathbox to main

This is in set.mm.  The immediate motivation is to use dvrdir in iset.mm
(combined with a desire to have theorems in iset.mm main also in set.mm
main).  But these do seem like generally applicable theorems for which
we already have the complex number analogues in main.

* Add dvrdir to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lringuplu to iset.mm

This is a key property of a local ring.

* Add ressmulrg to iset.mm

This is ressmulr from set.mm with an additional set existence condition.
The proof is similar.

* Add setscomd to iset.mm

Deduction form of setscom .

* Add mgpress to iset.mm

Stated as in set.mm.  The proof needs a lot of intuitionizing but is to
some extent based on the set.mm proof.

* Add rdivmuldivd to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a variety of
places but is basically the set.mm proof.

* Add lringring to iset.mm

A local ring is a ring.

* Add aprcotr to iset.mm

This is one of the theorems which shows that the ring apartness is an
apartness (although the ring has to be a local ring).

* Add lringnz to iset.mm

Being nonzero is one of the key properties of a local ring (as we
define it in df-lring anyway).

* Add aprap to iset.mm

This says that df-apr defines an apartness relation (for a local ring).

* Add local rings to set.mm

Copied without change from iset.mm

This is the syntax LRing , df-lring , lringring , lringnz , and lringuplu

* Comment changes mostly regarding apartness

This is in set.mm and iset.mm and contains various wording changes.
Also removes one unnecessary nesting of ${ $}

* Change TAp definition.

Instead of having df-tap be the one with the various properties expanded
and dftapap being the one in terms of Ap and proved in terms of that,
switch df-tap to be the one which had been dftapap and add dftap2 which
is the old df-tap, except proved from what is now df-tap .

* revise aprval comment