Add CCfld to iset.mm (#4678)
* copy plusgndxnn and basendxltplusgndx to iset.mm

Copied without change from set.mm

* Copy more slot theorems from set.mm to iset.mm

This is starvndxnbasendx , starvndxnplusgndx , and starvndxnmulrndx

* add grpstrndx to mmil.html

* add more scalar product slot theorems to iset.mm

Copied from set.mm.  They are vscandxnbasendx , vscandxnplusgndx ,
vscandxnmulrndx , and vscandxnscandx

* Add inner product slot theorems to iset.mm

Copied from set.mm.  This is ipndxnbasendx , ipndxnplusgndx ,
ipndxnmulrndx , and slotsdifipndx .

* Add le slot theorems to iset.mm

Copied from set.mm.

These are plendxnn , basendxltplendx , plendxnbasendx , plendxnplusgndx ,
plendxnmulrndx , plendxnscandx , plendxnvscandx , and slotsdifplendx

* Add UnifSet slot theorems to iset.mm

Copied without change from set.mm.

They are unifndx , unifid , unifndxnn , basendxltunifndx ,
unifndxnbasendx , unifndxntsetndx , and slotsdifunifndx

Update iset-discouraged

* Trim iset.mm comments

Specifically, remove references to theorems which iset.mm does not have
from various comments brought in with slot theorems.

* Add strext to iset.mm

Extending the upper range of a structure.

* Add CCfld to iset.mm

This is the syntax and df-icnfld .  df-icnfld is based on df-cnfld from
set.mm but with only the first 4 slots (at least for now), because
showing existence of the sets for the other slots depends on
theorems we haven't proved yet at this point in the file.

Add cnfldstr to iset.mm . Stated as in set.mm.  The proof is based on a
portion of the set.mm proof.

Update iset-discouraged file.

* copy cnfldex from set.mm to iset.mm

* Add cnfldbas to iset.mm

Stated as in set.mm.  The proof needs a bit of intuitionizing but is
basically the set.mm proof.

* copy addex and mulex from set.mm to iset.mm

* add cnfldadd to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add cnfldmul to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add cnfldcj to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Update cnfld entries in mmil.html

* Add more CCfld theorems to mmil.html

This is cnfldtset , cnfldle , cnfldds , cnfldunif , and
cnfldfun .