Intuitionize subgroups from df-subg to issubg2 (#4684)
* Fix comment to refer to SubGrp

There is nothing called SubGroup so refer to SubGrp instead which is
what must have been meant.

* Add ~QG , SubGrp , NrmSGrp to iset.mm

This is all three syntaxes and df-subg , df-nsg , and df-eqg

* Add issubg to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy subgss from set.mm to iset.mm

* Add grpressid to iset.mm

Restricting a group to its base set.

* Add subgid to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy subggrp from set.mm to iset.mm

* Add subgbas to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy subgrcl from set.mm to iset.mm

* Add subg0 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add subginv to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy subg0cl from set.mm to iset.mm

* copy subginvcl from set.mm to iset.mm

* Add subgcl to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy subgsubcl from set.mm to iset.mm

* Add subgsub to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy subgmulgcl from set.mm to iset.mm

* Add ifeq2dadc to iset.mm

This is ifeq2da from set.mm with a decidability condition added.  The
proof is based on the iset.mm proof of ifeq1dadc .

Correct comment for ifeq1dadc to reflect our conventions.

* Add subgmulg to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in several places
but is basically the set.mm proof.

* Add issubg2m to iset.mm

This is issubg2 from set.mm with non-empty changed to inhabited.
The proof needs intuitionizing but is basically the set.mm proof.