intuitionize subgroups from issubgrpd2 to trivsubgsnd (#4690)
* Add issubgrpd2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy issubgrpd from set.mm to iset.mm

* Add issubg3 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add issubg4m to iset.mm

This is issubg4 from set.mm with nonempty changed to inhabited.  The
proof needs intuitionizing at many steps but is basically the set.mm
proof.

* Add grpissubg to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy resgrpisgrp from set.mm to iset.mm

* copy subgsubm from set.mm to iset.mm

* Add subsubg to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add union/intersection theorems to mmil.html

These are intssuni , unissint , intssuni2 , uniintsn , and uniintab .
They had been mentioned in git logs but not in mmil.html until now.

* Add subgintm to iset.mm

This is subgint from set.mm with non-empty changed to inhabited.  The
proof needs intuitionizing in many places but is basically the set.mm
proof.

* Add 0subg to iset.mm

Stated as in set.mm.  The proof needs some intutionizing but is
basically the set.mm proof.

* copy trivsubgd from set.mm to iset.mm

* copy trivsubgsnd from set.mm to iset.mm