Intuitionize subrings (#4696)
* Add SubRing and RingSpan to iset.mm

This is the syntax, df-subrg , and df-rgspn .  Copied without change
from set.mm.

* Add issubrg to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy subrgss from set.mm to iset.mm

* Add ringressid to iset.mm

This is similar to grpressid but for rings.

* Add subrgid to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is similar
to the set.mm proof.

* copy subrgring from set.mm to iset.mm

* Add subrgcrng to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy subrgrcl from set.mm to iset.mm

* copy subrgsubg from set.mm to iset.mm

* copy subrg0 from set.mm to iset.mm

* copy subrg1cl from set.mm to iset.mm

* copy subrgbas from set.mm to iset.mm

* Add subrg1 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* copy subrgacl from set.mm to iset.mm

* Add subrgmcl to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add subrgsubm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add subrgdvds to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in several places
but is basically the set.mm proof.

* Add subrguss to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is basically the set.mm proof.

* Add subrginv to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add subrgdv to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in several places
but is basically the set.mm proof.

* Add subrgunit to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add subrgugrp to iset.mm

Stated as in set.mm.  The proof needs a fair bit of intuitionizing but
is basically the set.mm proof.

* Add issubrg2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add opprsubrg to mmil.html

* copy subrgnzr from set.mm to iset.mm

* Add subrgintm to iset.mm

This is subrgint from set.mm with non-empty changed to inhabited.  The
proof needs intuitionizing in various places but mostly follows the
set.mm proof.

* Add subrgin to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add subrgmre to mmil.html

Also, reword the entries for df-mre and df-mrc .  On the face of them
they aren't about closed sets defined as complements of open sets, so
may be more suitable for iset.mm than the previous text made it sound.

* Add issubdrg to mmil.html

* Add subsubrg to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* copy subsubrg2 from set.mm to iset.mm

* Add issubrg3 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add resrhm to mmil.html

* add resrhm2b to mmil.html

* add rhmeql to mmil.html

* add rhmghm to mmil.html

* add rnrhmsubrg to mmil.html

* add cntzsubr to mmil.html

* add pwsdiagrhm to mmil.html

* Add subrgpropd to iset.mm

Stated as in set.mm.  The proof needs intuitionizing on most lines,
although in ways that don't seriously affect the structure of the set.mm
proof.

* add rhmpropd to mmil.html