Intuitionize equivalence relation in a group generated by a subgroup (#4707)
* Add releqgg to iset.mm

This is releqg from set.mm with set exitence conditions added.  The
proof is a new one.

* Add eqgfval to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add brabv to iset.mm

Stated as in set.mm.  The set.mm proof is not intuitionistic but there
is an easy proof from relopab and brrelex12i .

* copy eqgval from set.mm to iset.mm

* Add eqger to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but other than that is the set.mm proof.

* Minimize eqger in set.mm

* copy eqglact from set.mm to iset.mm

* Add eqgid to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* Add eqgen to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy eqgcpbl from set.mm to iset.mm

* Add /s theorems to mmil.html

This is qusgrp , quseccl , qusadd , qus0 , qusinv , and qussub

* Add lagsubg2 and lagsubg to mmil.html