Add image structure to iset.mm (#4713)
* Add image structure to iset.mm

This is the syntax "s , and df-iimas (a modified form of df-imas from
set.mm).

* copy fexd from set.mm to iset.mm

* Add imasival to iset.mm

This is imasval from set.mm but with the reduction in the number of
slots found in df-iimas .  The proof also needs a lot of intuitionizing
for differences in set existence theorems.

* Add imasbas to iset.mm

Stated as in set.mm.  The proof is from imasival and requires a lot of
set existence steps not needed in the set.mm proof.

* Add imasplusg to iset.mm

Stated as in set.mm.  The proof is from imasival and has a lot of set
existence steps not needed in set.mm.

* Add imasex to iset.mm

This is a set existence theorem, not needed in set.mm because it has
ovex .  The proof is similar to some of the other image structure
proofs.

* Add imasmulr to iset.mm

Stated as in set.mm.  The proof is to some extent based on the set.mm
proof but needs a lot of set existence steps.

* copy f1ocpbllem from set.mm to iset.mm

* copy f1ocpbl from set.mm to iset.mm

* copy f1ovscpbl from set.mm to iset.mm

* copy f1olecpbl from set.mm to iset.mm

* copy iunssd from set.mm to iset.mm

* Add imasaddfnlemg to iset.mm

This is imasaddfnlem from set.mm with set existence conditions added.
The proof needs intuitionizing in various places but is basically the
set.mm proof.

Also add imasaddvallemg which is imasaddvallem from set.mm with the same
set existence conditions added and similar changes to the set.mm proof.

* Add imasaddflemg to iset.mm

This is imasaddflem from set.mm with set existence conditions.  The
changes to the proof are simply those needed to use the set existence
conditions in invoking imasaddfnlemg .

* Add imasaddfn to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add imasaddval to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add imasaddf to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add imasmulfn to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add imasmulval to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add imasmulf to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* Add image structure theorems to mmil.html

This is for imasvscafn , imasvscaval , imasvscaf , imasless ,
and imasleval , that is the theorems for the slots we have not
defined in iset.mm (at least yet).